requirements.txt
accelerate==0.21.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.32.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==1.4.8
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.15.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.15.4
flash-attn==2.6.3
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
func_timeout==4.3.5
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==4.15.0
gradio_client==0.8.1
griffe==0.45.3
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
iopath==0.1.10
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.7
lagent==0.2.2
latex2mathml==3.77.0
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmengine==0.10.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
narwhals==1.6.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.3.5
opencv-python==4.10.0.84
openpyxl==3.1.4
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
phx-class-registry==4.1.0
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
progressbar2==4.4.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
ruff==0.6.3
safetensors==0.4.4
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.38.4
streamlit==1.35.0
sty==1.0.6
svgwrite==1.4.3
sympy==1.13.2
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
timm==0.9.11
tinycss2==1.3.0
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==2.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
validators==0.28.3
visual-genome==1.1.1
watchdog==4.0.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.11
XlsxWriter==3.2.0
xtuner==0.1.11
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.11
zipp==3.20.1
